+++ src/Util.js
@@ -1,7 +1,10 @@
 const Sms77Client = require('sms77-client')
+const globalThis = require('globalthis')()
+if (!globalThis.fetch) globalThis.fetch = require('node-fetch')
 
 module.exports = class Util {
-    static initClient(apiKey) {
+    static initClient(RED, config) {
+        const apiKey = RED.nodes.getNode(config.config).credentials.apiKey
         return new Sms77Client(apiKey, 'node-red')
     }