+++ src/nodes/sms/sms.js
@@ -1,58 +1,65 @@
-const BaseNode = require('../../BaseMessageNode');
+const Util = require('../../Util')
+const NodeUtil = require('../../NodeUtil')
 
 module.exports = function (RED) {
-    'use strict';
+    'use strict'
 
-    class Sms77SmsNode {
-        constructor(config) {
-            BaseNode(this, RED, config);
-        }
+    function Sms77SmsNode(config) {
+        RED.nodes.createNode(this, config)
 
-        _onInput(msg, send, done) {
+        const node = this
+        const nodeUtil = new NodeUtil(node, config)
+        const client = Util.initClient(RED, config)
+
+        this.on('input', async function onInput(msg, send, done) {
+            if (!send) send = () => node.send.apply(node, [msg, send, done]) // If this is pre-1.0, 'send' will be undefined, so fallback to node.send
+
             const params = {
-                debug: 'true' === Sms77SmsNode.CFG.debug,
-                delay: this._emptyStringFallback('delay'),
-                details: 'true' === Sms77SmsNode.CFG.details,
-                flash: 'true' === Sms77SmsNode.CFG.flash,
-                foreign_id: this._emptyStringFallback('foreign_id'),
-                from: this._emptyStringFallback('from'),
-                json: 'true' === Sms77SmsNode.CFG.json,
-                label: this._emptyStringFallback('label'),
-                no_reload: 'true' === Sms77SmsNode.CFG.no_reload,
-                performance_tracking: 'true' === Sms77SmsNode.CFG.performance_tracking,
-                text: this._emptyStringFallback('message', msg.payload),
-                to: this._emptyStringFallback('recipients', msg.topic),
-                ttl: this._emptyStringFallback('ttl'),
-                udh: this._emptyStringFallback('udh'),
-                unicode: 'true' === Sms77SmsNode.CFG.unicode,
-                utf8: 'true' === Sms77SmsNode.CFG.utf8,
-            };
+                debug: 'true' === config.debug,
+                delay: nodeUtil.emptyStringFallback('delay'),
+                details: 'true' === config.details,
+                flash: 'true' === config.flash,
+                foreign_id: nodeUtil.emptyStringFallback('foreign_id'),
+                from: nodeUtil.emptyStringFallback('from'),
+                json: 'true' === config.json,
+                label: nodeUtil.emptyStringFallback('label'),
+                no_reload: 'true' === config.no_reload,
+                performance_tracking: 'true' === config.performance_tracking,
+                text: nodeUtil.emptyStringFallback('message', msg.payload),
+                to: nodeUtil.emptyStringFallback('recipients', msg.topic),
+                ttl: nodeUtil.emptyStringFallback('ttl'),
+                udh: nodeUtil.emptyStringFallback('udh'),
+                unicode: 'true' === config.unicode,
+                utf8: 'true' === config.utf8,
+            }
 
-            Sms77SmsNode.CLIENT.sms(params)
-                .then(response => {
-                    let code = response
-                    if (params.json) code = response.success
-                    else if (params.details) code = response.split('\n')[0]
+            try {
+                const response = await client.sms(params)
+                let code = response
+                if (params.json) code = response.success
+                else if (params.details) code = response.split('\n')[0]
 
-                    const succeeded = [100, 101].includes(Number(code));
+                const succeeded = [100, 101].includes(Number(code))
 
-                    if (!succeeded) return this._done(done, JSON.stringify(response), msg);
+                if (!succeeded) return nodeUtil.onDone(done, JSON.stringify(response), msg)
 
-                    let failed = 0;
-                    let sent = 1;
+                let failed = 0
+                let sent = 1
 
-                    if (params.json) {
-                        sent = 0
+                if (params.json) {
+                    sent = 0
 
-                        for (const msg of response.messages)
-                            true === msg.success ? sent++ : failed++;
-                    }
+                    for (const msg of response.messages)
+                        true === msg.success ? sent++ : failed++
+                }
 
-                    this._onSuccess(sent, failed, send, msg, response, done);
-                })
-                .catch(this._errorHandler(done, msg));
-        }
+                nodeUtil.status(`${sent} sent | ${failed} failed`)
+                nodeUtil.onSuccess(send, msg, response, done)
+            } catch (e) {
+                nodeUtil.errorHandler(done, msg)(e)
+            }
+        })
     }
 
-    RED.nodes.registerType('sms77-sms', Sms77SmsNode);
-};
+    RED.nodes.registerType('sms77-sms', Sms77SmsNode)
+}