+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+3.3.0
+==================
+* use empty string for unset parameters
+* fix wrong context when using multiple nodes of the same type
+* bump sms77-client to ^2.25.0
+
 3.2.0
 ==================
 * add missing translations